# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Panel API Documentation",
        "description": "API documentation for Payment Panel",
        "contact": {
            "email": "support@telemedia.pl"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/helpdesk": {
            "get": {
                "tags": [
                    "Helpdesk"
                ],
                "summary": "List all tickets",
                "operationId": "34f6aa54a1b164eb6a86ed63edb09ebc",
                "responses": {
                    "200": {
                        "description": "List of tickets"
                    }
                }
            },
            "post": {
                "tags": [
                    "Helpdesk"
                ],
                "summary": "Create new ticket",
                "operationId": "61f8089cf49f3e61340dc5af6dd6eb98",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket created"
                    }
                }
            }
        },
        "/api/helpdesk/{id}": {
            "get": {
                "tags": [
                    "Helpdesk"
                ],
                "summary": "Get ticket details",
                "operationId": "445ea29ea600464d0d41cd9463b6a35e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket details"
                    }
                }
            }
        },
        "/api/helpdesk/{id}/reply": {
            "post": {
                "tags": [
                    "Helpdesk"
                ],
                "summary": "Add reply to ticket",
                "operationId": "1c921dba873f4e8d86077b5030fba55e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reply added"
                    }
                }
            }
        },
        "/api/invoices": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "List all invoices",
                "operationId": "2e947f49543378e8d1ac24fc3ebcdcc7",
                "responses": {
                    "200": {
                        "description": "List of invoices"
                    }
                }
            }
        },
        "/api/invoices/generate": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Generate invoice for payment",
                "operationId": "a57d2a6bd294c8873303b817b14a2f18",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payment_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Invoice generated"
                    }
                }
            }
        },
        "/api/payments": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "List all payments",
                "operationId": "afcbc19a07bf33c2fcb73b5c438b476c",
                "responses": {
                    "200": {
                        "description": "List of payments"
                    }
                }
            },
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create payment",
                "operationId": "887e39c11a3a8f55e74ff6c6dcec0aaa",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "number"
                                    },
                                    "payment_type": {
                                        "type": "string",
                                        "enum": [
                                            "payu",
                                            "bank_transfer"
                                        ]
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment created"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List all users",
                "operationId": "566aed8efb436ca506e0ef51a032ecd9",
                "responses": {
                    "200": {
                        "description": "List of users"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "operationId": "62aa784a0f1c8427110490c3fed961a5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Helpdesk",
            "description": "Helpdesk"
        },
        {
            "name": "Invoices",
            "description": "Invoices"
        },
        {
            "name": "Payments",
            "description": "Payments"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}